+++ modules/canvaskit/npm_build/types/canvaskit-wasm-tests.ts
@@ -532,6 +532,10 @@ function paragraphBuilderTests(CK: CanvasKit, fontMgr?: FontMgr, paint?: Paint) 
             heightMultiplier: 1.5,
             forceStrutHeight: true,
         },
+        disableHinting: true,
+        heightMultiplier: 2.5,
+        textDirection: CK.TextDirection.LTR,
+        textHeightBehavior: CK.TextHeightBehavior.DisableFirstAscent
     });
     const blueText = new CK.TextStyle({ // $ExpectType TextStyle
         backgroundColor: CK.Color(234, 208, 232), // light pink