+++ modules/canvaskit/tests/paragraph.spec.js
@@ -846,6 +846,41 @@ describe('Paragraph Behavior', function() {
         builder.delete();
     });
 
+    gm('paragraph_mixed_text_height_behavior', (canvas) => {
+        const fontMgr = CanvasKit.FontMgr.FromData(notoSerifFontBuffer);
+        expect(fontMgr.countFamilies()).toEqual(1);
+        expect(fontMgr.getFamilyName(0)).toEqual('Noto Serif');
+        canvas.clear(CanvasKit.WHITE);
+        const paint = new CanvasKit.Paint();
+        paint.setColor(CanvasKit.RED);
+        paint.setStyle(CanvasKit.PaintStyle.Stroke);
+
+        const wrapTo = 220;
+        const behaviors = ["All", "DisableFirstAscent", "DisableLastDescent", "DisableAll"];
+
+        for (let i = 0; i < behaviors.length; i++) {
+            const style = new CanvasKit.ParagraphStyle({
+                textStyle: {
+                    color: CanvasKit.BLACK,
+                    fontFamilies: ['Noto Serif'],
+                    fontSize: 20,
+                    heightMultiplier: 3, // make the difference more obvious
+                },
+                textHeightBehavior: CanvasKit.TextHeightBehavior[behaviors[i]],
+            });
+            const builder = CanvasKit.ParagraphBuilder.Make(style, fontMgr);
+            builder.addText('Text height behavior\nof '+behaviors[i]);
+            const paragraph = builder.build();
+            paragraph.layout(wrapTo);
+            canvas.drawParagraph(paragraph, 0, 150 * i);
+            canvas.drawRect(CanvasKit.LTRBRect(0, 150 * i, wrapTo, 150 * i + 120), paint);
+            paragraph.delete();
+            builder.delete();
+        }
+        paint.delete();
+        fontMgr.delete();
+    });
+
     it('should not crash if we omit font family on pushed textStyle', () => {
         const surface = CanvasKit.MakeCanvasSurface('test');
         expect(surface).toBeTruthy('Could not make surface');