+++ modules/canvaskit/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    to `RuntimeEffect.makeShader` as floats (like all other uniforms), and will be converted to
    integers internally, to match the expectations of the shader.
  - Add 'halfLeading' to `TextStyle` and `StrutStyle`.
+ - `ParagraphStyle` now accepts textHeightBehavior.
 
 ### Removed
  - `Picture.saveAsFile()`, in favor of `Picture.serialize()` where clients can control how to